refactor: add ptrCast, fix bool sorting, and clean up UINT128 type and include headers

+++ src/common/arrow/arrow_row_batch.cpp
@@ -334,7 +334,7 @@ void ArrowRowBatch::templateCopyNonNullValue<LogicalTypeID::UUID>(ArrowVector* v
         std::memcpy(vector->data.data() + pos * valSize, &val, valSize);
     } else {
         auto offsets = (std::uint32_t*)vector->data.data();
-        auto str = UUID::toString(value.val.int128Val);
+        auto str = UUID::toString((uint128_t)value.val.int128Val);
         auto strLength = str.length();
         if (pos == 0) {
             offsets[pos] = 0;
